feat: Add parentElement option to enable scoping.
By adding parentElement we are able to init the plugin only for a certain scope of images. Something like initing a jquery plugin.
Added `opacities` key to supply an array of opacities for each color to transform to rgba.
Added `colorsToExtract` to tell the plugin how many colors to sample.

BREAKING CHANGE: `attachBoxShadowTo` is now `styleTarget`. `opacity` and `opacitySecondary` are deprecated and will be removed soon.